backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
aiofiles==23.2.0
pillow==10.1.0
pandas==2.1.3
scikit-learn==1.3.2
web3==6.11.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
